# Perf: openmp for calaulate_HR in deepks_lcao.cpp