### Perf Comparison
test case: [test-case.zip](https://github.com/user-attachments/files/19024259/test-case.zip)
The table below summarizes the time cost by the `calculate_HR` function.
||before|after|
|---|---|---|
|OMP=6 mpi=1|31.00s|6.13s|
|OMP=12 mpi=1|32.38s|3.82s|